Tweak prompts file again and reformat
