+++ api/package.json
@@ -22,11 +22,7 @@
     "access": "public"
   },
   "license": "MIT or Apache 2",
-  "keywords": [
-    "hono",
-    "local",
-    "debugging"
-  ],
+  "keywords": ["hono", "local", "debugging"],
   "bin": {
     "fpx": "bin/cli.js"
   },
